requirements.txt
torch>=1.10.0
torchvision>=0.11.0
segmentation-models-pytorch>=0.3.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
albumentations>=1.0.0
tqdm>=4.62.0
matplotlib>=3.4.0
